migrate old sc
